% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getAQDMavailable.R
\name{getAQDMavailable}
\alias{getAQDMavailable}
\title{getAQDMavailable}
\usage{
getAQDMavailable()
}
\value{
Returns \code{TRUE} if the service reports READY, otherwise returns \code{FALSE}.
}
\description{
A function to check the availability of EPA's AQDM web service
}
\details{
Makes use of AQDM's serviceAvailable service to check the status of AQDM.
}
\examples{
if(getAQDMavailable()) print("The service is available!")
}

