% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getAQDMdata.R
\name{getAQDMdata}
\alias{getAQDMdata}
\title{getAQDMdata}
\usage{
getAQDMdata(..., synchronous = FALSE)
}
\arguments{
\item{synchronous}{Should a synchronous request be sent (See details)}

\item{\dots}{named query parameters to be sent to AQDM.}
}
\value{
If \code{synchonous = TRUE} a \code{data.frame} is returned with the requested
data. If \code{synchonous = FALSE} an \code{AQDMrequest} object is returned.
The \code{AQDMrequest} object can be used with \code{\link{getAQDMrequest}} to
retrieve the data once it is ready.
}
\description{
Function for requesting monitoring data from U.S. EPA's Air Quality Data Mart.
}
\details{
\dots should be a set of name/value pairs where the names correspond to the
variables used by the AQDM web service.

If \code{synchronous = TRUE} then a synchronous request is sent to AQDM. This
will result in raw data being immediatiately returned to R. EPA is more strict
on the size of the request when using the synchonous service, so for larger
data pulls \code{synchonous = FALSE} may be required.

When \code{synchonous = FALSE} then an asynchronous request is sent tp AQDM.
This will create an object of class \code{AQDMrequest} being returned. This
object can then be used to retrieve the data from the server when it is ready.
}
\examples{
\dontrun{
 x <- getAQDMdata(user = "me@email.com", pw = "abc123", param = "44201",
                  state = 18, county = 89, bdate = "20140101",
                  edate = "20141231")
}
}

