% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getAQDMlist.R
\name{getAQDMlist}
\alias{getAQDMlist}
\title{getAQDMlist}
\usage{
getAQDMlist(name, ...)
}
\arguments{
\item{name}{The name of the variable for which you want the list of valid values.}

\item{\dots}{name/value pairs for subsetting the requested data. (See details)}
}
\value{
Returns a \code{data.frame} with 2 columns. The first column contains the valide code, the
 second column contains a description of that code.
}
\description{
Function for request a list of valid values for a given parameter from
U.S. EPA's Air Quality Data Mart.
}
\details{
If the variable requires other variables when querying for data you must provide
 those other variables to \code{getAQDMlist} as well. (e.g., you must provide a \code{state}
 code if \code{name = 'county'}.
}
\examples{
\dontrun{
 #Get a data.frame of counties in Indiana
 counties <- getAQDMList("county", state = "18")

 #Get a list of criteria pollutants
 crit <- getAQDMList("param", pc = "CRITERIA")
}
}

