% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getList.R
\name{getList}
\alias{getList}
\title{Get List of parameter codes}
\usage{
getList(name, criteria, format = "data.frame")
}
\arguments{
\item{name}{the name of the parameter to list}

\item{criteria}{a named list of query parametersto subset the named parameter}
}
\description{
Use this function to access the \code{list} endpoint of the air quality data
mart.
}

