\name{getParameters}
\alias{getParameters}
\title{
  Build Parameter List for U.S. EPA AQDM Data Pulls
}
\description{
  Opens a \code{tcltk} gui that allows the user to more easily build the parameter list necessary to make data requests using \code{rAQDM}.
}
\usage{
  getParameters()
}
\value{
  a \code{list} of named elements that correspond to the parameters of an AQDM data request.
}
\references{
  \href{http://www.epa.gov/airdata/toc.html}{Query Air Data (QAD) User's Guide}
}
\author{
  Eric Bailey
}

\seealso{
  \code{\link{requestAQDM}} \cr
  \code{\link{importAQDM}}
}
\examples{
#Coming Soon
}