% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProfile.R, R/getProfileNotify.R, R/getStatus.R
\name{getProfile}
\alias{getProfile}
\title{Get information about monitors synchronously}
\usage{
getProfile(requestId)

getProfile(requestId)

getProfile(requestId)
}
\arguments{
\item{requestId}{the request Id of a previously submitted notify job.}

\item{\dots}{named query parameters to be included in the query}

\item{\dots}{named query parameters to be included in the query}
}
\description{
Use this function to access the \code{profile} endpoint of the air quality
data mart

Use this function to access the \code{profileNotify} endpoint of the air quality
data mart

Use this function to access the \code{status} endpoint of the air quality
data mart
}

