% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRawData.R, R/getRawDataNotify.R
\name{getRawData}
\alias{getRawData}
\title{Return monitoring data synchronously}
\usage{
getRawData(...)

getRawData(...)
}
\arguments{
\item{\dots}{named query parameters to be included in the query}

\item{\dots}{named query parameters to be included in the query}
}
\description{
Use this function to access the \code{rawData} endpoint of the air quality
data mart.

Use this function to access the \code{rawDataNotify} endpoint of the air 
quality data mart.
}

