% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRetrieve.R
\name{getRetrieve}
\alias{getRetrieve}
\title{Retrieve a submitted job}
\usage{
getRetrieve(requestId)
}
\arguments{
\item{requestId}{the request Id of a previously submitted notify job.}
}
\description{
Use this function to access the \code{retrieve} endpoint of the air quality
data mart
}

