% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getServiceAvailable.R
\name{getServiceAvailable}
\alias{getServiceAvailable}
\title{Check status of AQDM}
\usage{
getServiceAvailable()
}
\description{
Use this function to access the \code{serviceAvailable} endpoint of the air quality
data mart
}

