% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRequest.R
\name{makeRequest}
\alias{makeRequest}
\title{Generic Request function for data mart}
\usage{
makeRequest(service, query, error = "warn")
}
\arguments{
\item{service}{The endpoint that is being queried}

\item{query}{a named list of query parameters to be passed to the service.

This will return the raw content of the response from the server so additional
processing will be required by the functions that wrap it. For example, the 
list endpoints return a tab-delimited file that will need to be converted to 
a data frame or a list.}
}
\description{
This function is used by all other request functions. It is an attempt to 
keep things DRY. If U.S. EPA changes their endpoint this is where it should 
be updated.
}

