% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/openAQDMgui.R
\name{openAQDMgui}
\alias{openAQDMgui}
\title{openAQDMgui}
\usage{
openAQDMgui()
}
\value{
The returned value depends on selections made by the user.

If request type is asynchronous
and the "Request Data" button is clicked, an \code{AQDMrequest} object will be returned (See
\code{\link{getAQDMrequest}}).

If request type is synchronous and the "Request Data" button is clicked,
the requested data.frame will be returned once it is returned from the server.
(currently disabled by EPA).

If the "Set Defaults" button in clicked, \code{NULL} will be returned invisibly, and
\code{\link{setAQDMdefaults}} will be run with the selected values.

If the "Copy Function" button is clicked, a character vector of length one will be returned
container the text of a function that would make the selected request.

If the dialog is canceled in any way, \code{NULL} will be returned invisibly.
}
\description{
Open a simple GUI to facilitate requesting data from U.S. EPA's Air Quality
Data Mart
}
\details{
\code{openAQDMgui} opens a dialog that the user can use to manipulate the variables
needed to request data from AQDM.
}
\examples{
\dontrun{
 x <- openAQDMgui()
}
}

