\name{raqdm-package}
\alias{raqdm-package}
\alias{raqdm}
\docType{package}
\title{
Access to U.S. EPA's Air Quality Data Mart
}
\description{
Provides a direct interface to U.S. EPA's Air Quality Data Mart through use of its Query Air Data (QAD) web query interface. Allows for manual data queries as well as GUI-directed data queries.
}
\details{
\tabular{ll}{
Package: \tab raqdm\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-11-22\cr
License: \tab What license is it under?\cr
}
  Main functionality it...
}
\author{
Eric Bailey
Maintainer: Eric Bailey <ebailey@idem.in.gov>
}
\references{
\href{http://www.epa.gov/airdata/toc.html}{Query Air Data (QAD) User's Guide}
}
\keyword{ package }
\examples{
#Coming Soon
}
