% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setAQDMuser.R
\name{setAQDMuser}
\alias{setAQDMuser}
\title{setAQDMuser}
\usage{
setAQDMuser(username, password, save = FALSE)
}
\arguments{
\item{username}{Your AQDM username (your email address)}

\item{password}{Your AQDM password}

\item{save}{\code{logical} Should the username/password be saved for future sessions?}
}
\value{
This function returns \code{TRUE} invisibly if there are no problems.
}
\description{
Convienince function for setting your username and password for accessing the
Air Quality Data Mart.
}
\details{
You can set user and pw from the \code{\link{setAQDMdefaults}} function also.
}
\examples{
setAQDMuser("fake@email.com", "pass123")
}
\seealso{
\href{http://www.epa.gov/airdata/toc.html}{Query Air Data (QAD) User's Guide}
}

